Fix overlapping text in gr-message
Previously, the amount of space the date took up on the screen was
estimated, but different types of formatting (time with AM/PM vs Date)
took up different amounts of space. Estimating was not working. Instead,
update the collapsed message to use flexbox. The message content
shrinks/expands when the width of the page changes, and will fit
regardless of the width of the date in a particular message. This change
also displays text content in a regular div while collpased rather than
gr-formatted-text so that it can be constrained to a single line.

Bug: Issue 4608
Change-Id: Id34b93c33b376d522f1d396a69133f854ad4dbf6

Former-commit-id: e47896578a56f03051a2b358fb0695bb62fe7bc1