Improve email footer format
Gerrit change emails include metadata to support mail filters and mail
search. In HTML emails, these are included in the email using a hidden
block to avoid visual clutter. However, with [1], it is possible for the
footer content to change from message to message in such a way that
Gmail quoting moves the delta text into a block that is not hidden.

Refactor the footer generation to embed each line in a hidden block. In
this way, the content remains hidden when it is changed, and mail
search/filters function as before. The list of footers is defined in the
email classes rather than the templates to simplify this embedding.

Whitespace around footers in HTML emails is partially preserved to
improve readability in the message source.

[1]  I42f3c4fa6fd19ceac137c99fad42952fcf7b7d6b

Bug: Issue 4936
Change-Id: Ic6713a587b14a6b0ecf658f4a047675afa97a854

Former-commit-id: deb7f68c4a5b44a8a885d95796df784bcd742a6e