ChangeIndexer: Reindex if stale after every index update
The index system is subject to a race condition when index updates
happen out of order with respect to their primary storage updates.
Previously, we had no way of detecting this case at all, and we
potentially stored stale data in the index indefinitely. This
particular issue was probably not very serious in practice (not that
we have any data on it), because if a change is busy enough to hit
this race condition, probably there is at least one non-racy write
somewhere in its future, so we would have indexed the right data at
some point.

Change-Id: I04ac36ea55af2cb1e1126365cc6b011e0fb8548c

Former-commit-id: 27ec2bad32b4cc81e3a7d0bdf2b502ab5930fa6b