Only show the 'Edit Config' button if refs/meta/config is visible
We already show the 'Edit Config' button on the project screen only
for project owners, however it seems to be possible to setup access
rights in such a way that project owners can't see the
refs/meta/config branch. In this case the user can't see changes for
the refs/meta/config branch and the behaviour of the 'Edit Config'
button is confusing. It lets the user click on 'Edit Config' which
creates a change for refs/meta/config but the change is not visible to
the user due to missing read permissions for refs/meta/config.

Fix this by showing the 'Edit config' button only for project owners
that can see the refs/meta/config branch.

Change-Id: Id52ec9a40443cae3b2609001c6317d54f0ff05b8
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 65edf0dc4d058c0e8273a86f6ef41ea5dc5546ba