Introduce gr-dropdown element
Creates a new element based on the gr-account-dropdown that can also be
used to replace the css-only version in the main header.  gr-dropdown
accepts two arrays as input-- "items" for links and "topContent" for
text to display above the links. TopContent was needed for the account
dropdown use case, as it displays name and email address. It also
accepts a horizontal-align parameter.  gr-account-dropdown is refactored
to use the generalized element.

Bug: Issue 4666
Change-Id: I65905b81504d1daa46e78312f576bf7482a4577c

Former-commit-id: f8fb4c982d2e45747f265cc7f70b2ac2f1927119