Add extension point to customize detailed commit message generation.
Plugins implementing new ChangeMessageModifier interface can modify
commit message when change is being submitted by Rebase Always and
Cherry Pick submit strategies as well as when change is queried with
COMMIT_FOOTERS option.

Intended usage is insertion of custom footer lines depending on the
state of the change and the destination branch.

Bug: Issue 5003
Change-Id: Ica56762381996db5d6f72b9e04f4588089039b44

Former-commit-id: 6fdc8ebf3d92faf897d20c6f0f486d93c9603d1c