Allow commitLink replacements overlap
If text replacements introduced by multiple commitLink rules overlaps,
they are all applied for cases when the new texts do not overlap.

For example, if two rules replace text in the same line, they are both
applied as long as they both change different parts of the line.

Following configuration makes possible for multiple comma-separated
issues in the same line to be linkified:

[commentlink "bug"]
	match = (BUG=\\s*)(\\d+)
	html = $1<a href=\"http://some.url/$2\">$2</a>

[commentlink "bug1"]
	match = (BUG=\\s*\\d+\\s*,\\s*)(\\d+)
	html = $1<a href=\"http://some.url/$2\">$2</a>

Bug: Issue 4416
Change-Id: Ie2354c943e6ddc4d5ad8328420c5746e57952d16

Former-commit-id: fb48902ef303d008c6f0dd38d654106c00313530