Refactor PostReview#checkComments
In order to improve readability, the method PostReview#checkComments
is restructured into logical units. In addition, the code which is
responsible for cleaning up the inputs and the one which enforces
constraints on them are separated from each other. Thus, those two
different aspects are easily visible and the code complies more with
clean code principles (e.g. "Do one thing"). Because of the separation,
we now need to iterate twice over the comments.

Change-Id: Ia2d70588a4b1d6b8fc08743c0afa67f7a2fffb45

Former-commit-id: cab73be56c07d10165d9a0f4e816efd8fda1beed