Suppress Edit Description button in the GWT UI
Exposing it as a non-core action is confusing, since it produces a
button labeled "Edit Description" that throws an exception and
freezes the UI instead of editing a revision's description.

Apparently actions like this that require input need to be either
core actions or handled by a plugin. This one is a core action.

Bug: Issue 5019
Change-Id: I8ab4f527c25009df93328a6c505a4184d790ace9

Former-commit-id: 5fd562e784493956c57590ef6eb164b816cd62f8