Check permissions on refs/meta/config on save of ACLs for review
Users should not be able to create a change for a branch that is not
visible to them, especially since the created change would be not
visible to them.

Creation of changes for the refs/meta/config in addition requires that
the user has the permission to upload changes to this branch or is
project owner.

If a project owner tried to save access rights modifications for
review it failed with a Server Error showing a NotFoundException,
since the created change was not visible to the user. Now the creation
of the change fails and a proper error message is displayed.

Change-Id: I6ca243eb8546011427c921e6e1fdef23626f0fa6
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 9023196e86f6856bf6e99565b131dce12daaf125