Make BranchResource and TagResource extend a common class
Add a new abstract class, RefResource, that defines access methods to
get the ref name and revision.

Make both BranchResource and TagResource derive from RefResource, and
add the missing implementations in TagResource.

This will make it easier to consolidate common code that can be used
for both branch and tag related endpoints.

Change-Id: Ib44d621e34ce2b1077349db0037d960155b790e0

Former-commit-id: eff1643099f0cc373ab620deb95b6da24b91b205