WebLinks: Return List<T> from methods rather than FluentIterable<T>
All the callers of these methods convert the returned FluentIterable to
a List. Also it's inefficient to invoke FluentIterable.isEmpty() because
it has to traverse the whole thing including map and filter calls.

Change it to just return a List.

Rather than converting to use the Java 8 streams API, leave the internal
implementation of the methods unchanged for now. The links are DynamicSet,
which is an implementation of Iterable and does not have the stream()
method. There isn't much readability improvement by rewriting it to use
Spliterator, but we can revisit this later when we upgrade to Guava 21
which has a Streams utility.

Change-Id: I0e2a68f2992f6bcb2ab1ee2ae067975d416e9651

Former-commit-id: e32ef82ad739cbb0c7ba775d4d37f21599eee816