Do not assume DB down if change to index not found
The Gerrit SSH command "index changes" may receive change numbers that
cannot be found or refer to repositories that have problems.

The assumption of "DB is down" when receiving an OrmException was wrong
because the JGit errors were simply wrapped as OrmException and
do not necessarily mean that you cannot continue to process the other
ones in the list.

Change-Id: I1098be16b682dce8d4e63120e7e50f937b19a7b8

Former-commit-id: 9a52dd1dce8a79b90e8bfad4f718659aafe1a3e8