Add support for parent revisions weblinks provided by plugins
Support for parent weblinks was removed in change I8508dcaf3. This is a
regression for some users who rely on that information being present.

On the other hand, not all users want to have these links going to the
same external service as the patch set revision links, so we can't just
add it back as it was.

Instead, add a new plugin web link interface for the parent revisions.

If a plugin provides an implementation of ParentWebLink, the links will
be displayed beside the parent revisions on the change screen.

For the built-in gitweb support, provide the parent links using the same
template as the patch set links, so they will always appear when the
gitweb config defines the 'revision' setting.

Bug: Issue 4908
Change-Id: Icdf0f997f939128717ee4edc2c147a26eeaebad6
(cherry picked from commit 3e305750440647030d6cd85bb84b24cea554d8a7 [formerly 58b8d76204ec59fa2550f450921a34bac2f0cea6])

Former-commit-id: 1aedee62aeb4d9ca559f70ca3efc4186d93fb58c