Revert "AccountManager: Check that ext ID belongs to account before delete"
I think this change didn't make sense. The external ID that is deleted
is initialized with the account ID of the new account, hence the check
added by this change would always match. If that external ID belonged
to another account it would have been overwritten when the external ID
was written with 'upsert' to the database.

This reverts commit b25f0a3b06e1918cc3ca5239821a2537c8f86ab0 [formerly 352be569f969c9cd16032c911365882779c22ce4].

Change-Id: Ibbbcaa77767ad06cef976a3236abc64ed1d8b71d

Former-commit-id: 5862be62748a552ad9f68b3f021451c81f7336b4