Have PolyGerrit footer link preserve current page
The previous implementation used getHostPageBaseUrl, which seems
logical, but because Gerrit is a GWT app that uses fragment urls
to represent logical pages, this results in always using '/' as
the host page base url. This new implementation instead uses the
history token, which is the fragment url minus '#', as input to
a full UrlBuilder, and ensures it converts both the fragment to
a path, and the @-style pseudofragment to an actual fragment.

Change-Id: I09b8a711a8d56d36c4c3b33630c456a84097fe37

Former-commit-id: 2981ff7c9a945e6e605db7757afa30f3d7b90bb6