Bazel: Expose neverlink artifact in maven_jar rule
Given that the feature request: [1], still wasn't implemented, expose
neverlink artifact for every artifact per default. This is needed to
support in Gerrit tree build for plugins, because every dependency
can be potentially used from a plugin, in which case it must be used
in neverlink form to avoid shipping it twice, in plugin artifact in
addition to gerrit.war itself.

With this change in place, we can write this build rule:

  gerrit_plugin(
      name = "verify-status",
      [...]
      provided_deps = [
          "@commons_dbcp//jar:neverlink",
      ],
      [...]

that would now work in both build mode: standalone (using bazlets) and
in Gerrit tree mode.

[1] https://github.com/bazelbuild/bazel/issues/1402

Change-Id: I1240d25c576b13bd4d7450a0e5ba143df27a3d3a

Former-commit-id: c13a3db81ad01ed9c96bb03779c90fe82d118e2d