Highlight label votes in HTML email
When a comment email includes label votes, they can be difficult to find
or hard to interpret. For example, the semantic distinction between
Code-Review: +2 and Code-Review: -2 is significant, but in email are
typographically similar.

Votes were included in emails as part of the comment "message". which is
compiled by com.google.gerrit.server.change.PostReview#insertMessage and
would follow the following format:

  Patch Set [# of patch set]: [votes...]

  [top-level comment]

  ([# of comments] comments)

To highlight votes in email, the list of votes (the vote delta) and the
top-level comment are passed to the email template as separate
parameters. In HTML emails, the comment message is built using these
parameters with additional formatting applied to votes.

Plain-text emails are unchanged and include the message text as before.

Feature: Issue 4744
Change-Id: Ia5430edaeb2777c6afdcc56617bcc90a0eb196d1

Former-commit-id: dc05dce34088f5f75ebccebd40f15770d7346fdd