ChangeNotesParser: Don't sort PatchSets
This was only necessary to get the latest patch set, which was then
used as the current patch set. Since the current patch set is no
longer set this way, sorting is no longer necessary. The only caller
of ChangeNotesState#patchSets already re-sorts the output, so no need
to do it twice.

Change-Id: Iba0afc84f45a45b25d08a29b87fc7eb87ad49a72

Former-commit-id: 223dce63b1efcb6265606d24cddc1c2a9eea5fd7