Implement NoteDB portion of resolvablility
The server doesn't need to handle resolvedness of a thread; the plan is
for it to be an entirely client-side implementation. Whatever value the
most recent comment has for 'unresolved' is the value that will
determine whether the thread is resolved, but each individual comment
will still have that value populated so as to keep a log of when it was
changed and who it was changed by.

Change-Id: If45057935f4f30bb6babbe5e80223b8da3fa694f

Former-commit-id: c101b42cc3ca7e0a6bd810a6f589e44f8d19dfad