Remove threading in comment UI
Sort inline comments in the order that the comments were written.
Formerly comments were sorted in tree order so that reply comments would
appear after their parent. However, this sorting could be confusing
because there would be no other indication that the comments were
associated to a parent. When comments are sorted chronologically, they
appear in the thread position a user would expect.

Change-Id: I4d0722e4511eb53b9b8ab1d61a922d1b142e09ad

Former-commit-id: cc518249be51af7a4b1a37ff36536b7a7d9080e0