Initialize download scheme preference
The download dialog checks the logged-in property to determine whether
it should check user preferences for a default download scheme.
Previously, however, it would check the flag before it had been
completely initialized (it starts out with a value of false) resulting
in the preferred scheme never being selected.

With this change, the download dialog reacts to changes in the logged-in
property instead of in the "attached" lifecycle method.

Bug: Issue 4203
But: Issue 5147
Change-Id: Ic412b1d9a04cc6083b090d7767172e541578e5c3

Former-commit-id: 2345a5d19c4b1715f7f4283bcc75be0935597142