elasticsearch: Always set up both change and account indexes
Change query tests subtly depend on the account index for the account
setup step: omitting the account index caused
AccountManager#authenticate to throw an exception if an account
external ID was repeated, instead of returning the account with that
ID.

Always set up all indexes, using a helper in ElasticTestUtils. The
delete code can be simplified as well using the special string "_all"
to delete all indexes.

Change-Id: I7ab7be37a8cf953b6873745014e781e6b5e910f8

Former-commit-id: 49251588f2379381ea4b7367a6002f9099952bcc