Make the file path for the replacement of a robot comment explicit
Up to now, the file path was implicitly limited to the file on which
the robot comment was placed. This restriction will be relaxed in the
future. In order to not break systems which use the fix suggestions at
that (later) point in time, the mandatory field 'path' is introduced
with this change.

Change-Id: I73875b97bf844dee927720765046161bbd888c02

Former-commit-id: 110f60f08226facd717683f045753e8ef58a2344