Don't require Add Patch Set permission for submit by rebase
When the submit strategy was Rebase Always or Rebase If Necessary and
a rebase was needed for the submit, the submit failed if the user
didn't have the Add Patch Set permission. However for submitting a
change the Submit permission alone should be sufficient.

The behavior is now consistent with the Cherry-Pick submit strategy
which also doesn't require the Add Patch Set permission if a
cherry-pick is done on submit.

Change-Id: Id9c484ff51f9dfa7ea77216fbf9b06a799676412
Signed-off-by: Edwin Kempin <ekempin@google.com>
(cherry picked from commit 561cd017a3236122721dbcc410e41e8b40ff7599 [formerly 969a1953cf3889baa663fe59e421ba9cffcbabe3])

Former-commit-id: 43fe972939560029f5918bd50ad3e3e287c48620