Improve documentation of diff-match-patch build
Output the actual jar path used for the build in the generated JS
file, rather than a broken Google Code link. This requires extracting
a constant for the version string.

Rename the addon_merge target to indicate that it includes bundled
dependency code from diff-match-patch, which fact is not otherwise
obvious.

Change-Id: I6512e5d65b81802d917f718f8e62ac414e8bffda

Former-commit-id: 8cadb33267037ef4052b7bc9e8376e56ad894f92