Set server_tests size to large for CI
//gerrit-server:server_tests takes on average 2' on a local
multi-cpu machine. On the CI, during peak usage, it may exceed the limit
of 5' (300 sec) for defining a failed test suite on Bazel.

As a matter of fact, this suite is stable on Buck but mostly
failing for Bazel.

Raise the limit to large to make it more stable on Bazel
builds on the CI.

Change-Id: I20a02d4750d929bbe4e40bbf7a46356051cd4c58

Former-commit-id: 08223f8bb6c5b20da590954c1fed42312d222a5f