Add new optional interface for REST collections to get request params
REST collections decide which REST endpoint is used to handle GET
requests on the collection by returning the implementation of this
REST endpoint from the list() method. Sometimes depending on the
presence of certain URL parameters it makes sense to let different
REST endpoints handle the listing. For this the REST collection needs
to know which URL parameters were specified. By implementing the new
NeedsParams interface the REST collection can now get to know the URL
parameters.

Change-Id: I78a7b7a26324c63bc7ee24e7f0928f907146981d
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: d819316d4255ecb8cf061607eba8eedd9854fb78