Resolve username lazily in git over http tasks
ProjectQoSFilter uses username as a part of task info. However,
AuthFilter which comes later than ProjectQoSFilter renders the
value of username. As a result, username never gets resolved when
using git with http.

The current patch set resolves the username only when it is used,
i.e., in the toString method, and saves the value for reusage.

Bug: Issue 5206
Change-Id: I280c95b242d599919eee3c7007c60a0047170166

Former-commit-id: 3fd92a4ac50e5ef49043858ab2576050634a3c2e