ChangeData: Filter out zombie draft refs
Since I622dbbb3 the implementation of ChangeData#draftsByUser just
took the set of account IDs that have corresponding draft refs.
However, this is not actually the same as the set of users who have
drafts, because some drafts may be zombies: drafts where the published
comment was added to the change meta ref, but the later write to
All-Users to delete the draft failed. Since there is no atomicity
between the change repo and the All-Users repo, this can be expected
from time to time in normal operation. For this reason,
ChangeNotes#getDraftComments already knows how to filter out zombies.

Tweak the implementation of ChangeData#draftRefs to filter out refs
where all the comments on the ref are zombies. This restores the old
behavior of draftsByUser, which fixes the draftby search operator,
which is the main thing I'm concerned about. This is not strictly
correct because we will still return an over-populated ref from
draftRefs in the case where some but not all comments are zombies.
However, trying harder to fix this corner case is not worth the
effort. First of all, this case should be quite rare in the first
place, because the default behavior of PostReview is to publish all
comments on all revisions, deleting the ref. Second of all, as
mentioned, the main thing I'm concerned about is the set of users who
have drafts, and this implementation preserves that.

Change the implementation of ChangeNotes#getDraftComments to cache the
set of Comment.Keys in the published map. The linear search was more
ok when we weren't calling getDraftComments in a loop over all users.
Plus this new implementation results in slightly cleaner code.

Change-Id: I93ac69db418a8ad068d1919fd6e0fb39d290df5f

Former-commit-id: 9d0236f88e235d569c1b59eb80c77807696b6d50