Add keyboard shortcuts for expanding and collapsing all messages
This change adds keyboard shortcuts for expanding and collapsing all
messages on change views for feature parody with the GWT UI.

The 'x' key expands all comments, and the 'z' key collapses all
comments.  In addition, the expand all/collapse all text on toggle
button changes when the keyboard shortcuts are pressed (i.e. when x is
clicked and all messages expand, the text will change to 'Collapse
All').

Bug: Issue 5193
Change-Id: Ia60abe22eec3ef9b4ab5baebdb74d13f3efa8eeb

Former-commit-id: 6bc285d79059ca18b988ee1e1337259b3df4b0a3